+++ src/chain/tezos/TezosMessageUtil.ts
@@ -1,14 +1,9 @@
 /*eslint no-bitwise: 0*/
 import base58check from "bs58check";
-import baseN from "base-n";
 import bigInt from 'big-integer';
 
 import { SignedOperationGroup } from '../../types/tezos/TezosChainTypes';
 import { CryptoUtils } from '../../utils/CryptoUtils';
-
-const base128 = baseN.create({
-    characters: [...Array(128).keys()].map(k => ('0' + k.toString(16)).slice(-2))
-});
 
 /**
  * A collection of functions to encode and decode various Tezos P2P message components like amounts, addresses, hashes, etc.
@@ -37,7 +32,7 @@ export namespace TezosMessageUtils {
     export function writeInt(value: number): string {
         if (value < 0) { throw new Error('Use writeSignedInt to encode negative numbers'); }
         //@ts-ignore
-        return Buffer.from(Buffer.from(base128.encode(value), 'hex').map((v, i) => { return i === 0 ? v : v ^ 0x80; }).reverse()).toString('hex');
+        return Buffer.from(Buffer.from(CryptoUtils.twoByteHex(value), 'hex').map((v, i) => { return i === 0 ? v : v ^ 0x80; }).reverse()).toString('hex');
     }
 
     /**
@@ -81,10 +76,9 @@ export namespace TezosMessageUtils {
      * @param {string} hex Encoded message part.
      */
     export function readInt(hex: string): number {
-        return base128.decode(
-            //@ts-ignore
-            Buffer.from(Buffer.from(hex, 'hex').reverse().map((v, i) => { return i === 0 ? v : v & 0x7f; })).toString('hex')
-        );
+        //@ts-ignore
+        const h = Buffer.from(Buffer.from(hex, 'hex').reverse().map((v, i) => { return i === 0 ? v : v & 0x7f; })).toString('hex')
+        return CryptoUtils.fromByteHex(h);
     }
 
     export function readSignedInt(hex: string): number {